+++ CHANGELOG.md
@@ -10,7 +10,9 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - added support for NeoForge test framework
 - added support for JUnit tests
+- added global option to change mod package
 - changed test mod definition to decouple it from vanilla game tests
+- changed build config feature to use global mod package path
 - updated internal ModDevGradle version to 2.0.138
 - updated minimum ModDevGradle version to 2.0.138 for compatibility with unobfuscated sources
 - updated default Java target version to Java 25